+++ starkbank/__init__.py
@@ -1,4 +1,4 @@
-version = "2.10.1"
+version = "2.11.0"
 
 user = None
 language = "en-US"