+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- "payment" account type for Pix related resources
+
 ## [2.10.1] - 2021-06-07
 ### Fixed
 - TaxPayment.scheduled being cast to datetime instead of date